avoid direct freetype usage in core/public hpp file so that mapnik C++ depedencies do no need to known the freetype include paths
