+++ src/grid/process_text_symbolizer.cpp
@@ -24,6 +24,7 @@
 #include <mapnik/feature.hpp>
 #include <mapnik/grid/grid_renderer.hpp>
 #include <mapnik/symbolizer_helpers.hpp>
+#include <mapnik/font_util.hpp>
 
 namespace mapnik {
